+++ CHANGELOG.md
@@ -188,6 +188,8 @@ They also can only drop from certain mobs, for example bonechill ring can only d
 * instead of item stat ranges, you now get a "[x%]" percentage view of the stat, so you can easily know if it's a good roll or bad. Brought to you by popular bad guy color, red! And of course, the purple king!
 * optscaleexactstat field names changed, possibly datapacks changed
 * removed world areas mechanic, wasn't used really ever since it was invisible
+* new datapack type: races. They contain extra stat scalings (like being able to gain mana with more dexterity), base stats, etc.
+* unique gear datapacks now need another field "gear_types", this determines which gear types can get this unique. 
 
 * salvaging is now a button click like how modifying/socketing is now too
 * added "secret" salvaging recipes (datapacks can't keep secrets though)